Refactored GatekeeperClient
